Initial looping to go through matrices diagonally
Added a first attempt at a loop to work through the HMM matrices
diagonally. I am not working with arrays yet. This should go through any
given [rows][columns] matrix and access the cells in diagonal order.

I have not actually run this yet in the context of the HMM matrices. Not
sure if it will work.

Probably broken- the initial row,column values (not taking hapStartIndex
into account)